travis: Build libtpms for coverage and test with swtpm
Build libtpms with code coverage enabled and run the swtpm test
suite on it.

Unfortunately cpp-coveralls has a problem creating correct file paths when
collecting the results so that the coverage results are not correctly sub-
mitted to coveralls.io. The issue may stem from the fact that the
Makefile.am in src/ contains rules for source files in subdirs src/tmp12/
and src/tpm2/ etc. I hope that this issue can be fixed at some point.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>