tpm2: Add COMPAT_LEVEL to permanent state to track compatibility level
Add a variable compatLevel to the each seed in the permanent state
that allows us to track the age of the seed. Whenever a new seed
is created the compatLevel is also written and set to the latest
version. This compatLevel then influences the crypto algorithm that
can be used.

This patch only sets the variables to the current compatibility
level COMPAT_LEVEL_ORIGINAL and writes it out as part of the state
file. This makes the state file not downgradeable.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>