tpm2: Introduce COMPAT_LEVEL_OPENSSL but do not use yet
Introduce the COMPAT_LEVEL_OPENSSL constant so that we can add the
code that will be execute once this compatibility level is activated.

The meaning of this constant is that all derived keys are now generated
by OpenSSL. We will activate this by setting COMPAT_LEVEL_LAST to
COMPAT_LEVEL_OPENSSL but can only do this once all code has been added to
create derived keys with OpenSSL API calls. (OpenSSL will need to be extended).

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>