tpm2: Sanity check an OBJECT loaded via context
Perform some basic sanity checks on an OBJECT loaded via context to make
sure that no reserved flags are set or buffers have an invalid size. The
latter could occur if the context was created on a system with a different
endianess, which would be possible with libtpms's state portability.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>