tpm2/build-sys: Add PC Client compliant tpm2
Most TPMs in consumer PCs are compliant to the TCG PC Client Platform
TPM Profile (PTP) Specification. To addtionally build a library
libtpms-pcclient with minimal PC Client feature set, the configure
option --with-extra-profiles=pcclient is added.

The values specified in Implementation_pcclient.h are based on the spec
version 01.04rev37. By default, libtpms-pcclient is NOT built.

Signed-off-by: Johannes Holland <johannes.holland@infineon.com>