nvIndex: Fix null pointer dereference
Variable nvIndex is obtained via calling NvGetIndexInfo.
Unfortunately, the function call can return a NULL
pointer if the index was not found.

After the nvIndex is populated, potentially with NULL,
the pointer is accessed without checks. There are
places where this check happens, but they are
not exhaustive.

Ensure that all nvIndex obtained via NvGetIndexInfo
are guarded with asserts to avoid access to invalid
memory regions.

Signed-off-by: Alexandru Vasile <lexnv@amazon.com>