tpm2: Filter-out disabled commands from being reported or executable
Filter-out disabled commands so that the GetCapability() command does not
return them in the list of implemented/supported commands. Also prevent
them from being executable. Typically the check for whether a command
is enabled/disabled needs to be added around checks for whether the
IS_IMPLEMENTED flag is set on a command.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>