tpm2: Add TPMT_PUBLIC & TPMT_SENSITIVE params to CryptRsaLoadPrivateExponent
In the next patch there will be ObjectLoad calling
CryptRsaLoadPrivateExponent without providing an OBJECT to this function
but only to use it to check the consistency of TPMT_PUBLIC and
TPMT_SENSITIVE parameters. To allow this, add these two parameter to the
function (= sync with upstream) but also keep the rsaKey as an optional
parameter (libtpms only) so that the flag attributes.privateExp can still
be set on the key in case the values for dP, dQ, qINv are valid.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>