tpm2: Comment and rearrange some code
Rearrange the call to RsaInitializeExponentOld in CryptyRsaGenerateKey
so that this function can be removed from the OpenSSLCryptRsaGenerateKey
implementations.

Sync the PackExponent and UnpackExponent functions calls with upstream
but comment them since libtpms does not used them (will not remove
privateExponent from OBJECT).

Add comments to diffs between libtpms and upstream.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>