rev180: Add TPM2_DefineSpace2 and dependencies
TPM2_DefineSpace2 defines NV areas in the same format as the old
TPM2_DefineSpace function did, so there should not be any compatibility
issues with the NV areas' format. However, it may not currently make
much sense to enable this command since its improvement over the existing
TPM2_DefineSpace is only support for exernal memory, which we don't have.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>