tpm2: Use marshalling to write larger OBJECTs into NVRAM
Instead of copying an OBJECT directly from memory into NVRAM memory
determine the type of OBJECT (RSA vs. ECC etc.) and marshal the OBJECT
as ANY_OBJECT into a buffer and copy the buffer into NVRAM. To maintain
backwards compatibility copy RSA keys with size 3072 and smaller
directly into NVRAM (OBJECT is copied as before).

When reading an OBJECT out of NVRAM try to unmarshal it as an
ANY_OBJECT first and if this does not work it is with great likelyhood
an RSA3072_OBJECT that needs to be copied from NVRAM and than
transformed into an OBJECT.

The new marshalling of OBJECTS will only be used once RSA 4096 keys
are enabled or other changes to the code are made.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>