tpm2: NVMarshal: Add missing case of skipping a block when none is there
Add handling of the case of wanting to skip a block of code when no data
are in the byte stream. This case has not occurred so far where a block
of unmarshalling code needed to be skipped but also no data were there in
the byte stream - it would have otherwise lead to errors while trying to
unmarshal data that were not there. So far there was simply no code there
that should have been skipped.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>